Refactored pages module. Part of codebase cleanup effort
