+++ gui/pages/__init__.py
@@ -1,0 +1,24 @@
+from .start import StartPage
+from .project_select import SelectProjectPage
+from .project_new import NewProjectPage
+from .importer import ImportDatasetPage
+from .analyzer_select import SelectAnalyzerForkPage
+from .analyzer_new import SelectNewAnalyzerPage
+from .analyzer_previous import SelectPreviousAnalyzerPage
+from .analysis_configure import ConfigureAnalysis
+from .analysis_params import ConfigureAnalaysisParams
+from .dataset_preview import PreviewDatasetPage
+
+
+__all__ = [
+    "StartPage",
+    "SelectProjectPage",
+    "NewProjectPage",
+    "ImportDatasetPage",
+    "SelectAnalyzerForkPage",
+    "SelectNewAnalyzerPage",
+    "SelectPreviousAnalyzerPage",
+    "ConfigureAnalysis",
+    "ConfigureAnalaysisParams",
+    "PreviewDatasetPage",
+]
